Fix bug preventing processing of last sample
The previous version duplicated the last row of the table instead of processing
the last sample. This was due to a missing character on the line computing the
progress bar. As a consequence, the last sample name was not extracted, and the
second-to-last sample name was used instead (hence the repeated line). This
means that the last sample was actually never processed.